// src/typings/model.ts
export type EdgeConfigs = {
  /**
   * Edge type, which can be a built-in edge or a custom edge.
   * 连线类型，可以是内置连线也可以是自定义连线
   */
  type?: string | number;
  /**
   * Uuid of the edge.
   * 连线唯一标识
   */
  id?: string;
  /**
   * Source node id of the edge.
   * 源节点 id
   */
  source?: string;
  /**
   * Target node id of the edge.
   * 目标节点 id
   */
  target?: string;
  /**
   * Customize the start coordinates.
   * 自定义起始点位置
   */
  startPoint?: number[];
  /**
   * Customize the end coordinates.
   * 自定义终止点位置
   */
  endPoint?: number[];
  /**
   * Customize the end coordinates.
   * 连线文本标签配置
   */
  label?:
    | string
    | (LabelConfigs & {
        /**
         * Percentage positioning of the label relative to the edge.
         * 文本标签相对于连线的百分比定位，默认为 `0.5`
         */
        position?: number;
      });
  /**
   * Customize the loop edge styles.
   * 自环连线配置
   */
  loop?: LoopConfigs;
  /**
   * The control points of the edge.
   * 连线控制点
   */
  controlPoints?: number[][];
  /**
   * The styles of curve or turning line.
   * 曲线样式
   */
  styles?: {
    /**
     * Percentage positioning of the turing point relative to the edge.
     * 拐点相对于连线的百分比定位
     */
    curvePosition?: number | number[];
    /**
     * Absolute offset of the turning point.
     * 图标相对于节点的绝对偏移(px)
     */
    curveOffset?: number | number[];
    /**
     * The radius of the turning angle.
     * 拐角圆角大小
     */
    radius?: number;
  };
  /**
   * The anchor index of source node.
   * 指定连接到源节点的一个锚点
   */
  sourceAnchor?: number;
  /**
   * The anchor index of target node.
   * 指定连接到目标节点的一个锚点
   */
  targetAnchor?: number;
  [key: string]: any;
} & BaseConfigs;
// ...
export type LabelConfigs = {
  /**
   * Content of the label.
   * 文本标签内容
   */
  text?: string;
  /**
   * Absolute horizontal offset of the label(px).
   * 文本标签内容的绝对水平偏移(px)
   */
  offsetX?: number;
  /**
   * Absolute vertical offset of the label(px).
   * 文本标签内容的绝对垂直偏移(px)
   */
  offsetY?: number;
  /**
   * The background rect shape configuration of the label.
   * 文本标签背景矩形样式配置
   */
  background?: {
    /**
     * The padding between rect and label.
     * 背景矩形与标签文本之间的留白
     */
    padding?: number[];
    /**
     * The fill color of the background.
     * 背景矩形的填充色
     */
    fillStyle?: string;
    /**
     * The stroke color of the background.
     * 背景矩形的边框色
     */
    strokeStyle?: string;
    [key: string]: any;
  };
  /**
   * Text horizontal alignment.
   * 文本水平对齐方式
   */
  textAlign?: string;
  /**
   * Text vertical alignment.
   * 文本垂直对齐方式
   */
  textBaseline?: string;
  /**
   * Text font weight.
   * 文本字体粗细
   */
  fontWeight?: number;
  /**
   * Text font size.
   * 文本字号
   */
  fontSize?: number;
  /**
   * Text font family.
   * 文本字体
   */
  fontFamily?: string;
  /**
   * Text line height.
   * 文本行高
   */
  lineHeight?: number;
} & BaseConfigs;
// ...
export type LoopConfigs = {
  dist: number;
  position: 'top' | 'left' | 'bottom' | 'right';
  clockWise: boolean;
};